MDL-72509 core: Move participants filter from core_user to core
