--- user/amd/src/local/participantsfilter/filtertypes/country.js
@@ -1,36 +1,0 @@
-// This file is part of Moodle - http://moodle.org/
-//
-// Moodle is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// Moodle is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-// GNU General Public License for more details.
-//
-// You should have received a copy of the GNU General Public License
-// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-
-/**
- * Country filter
- *
- * @module      core_user/local/participantsfilter/filtertypes/country
- * @copyright   2021 Paul Holden <paulh@moodle.com>
- * @license     http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
- */
-
-import Filter from '../filter';
-
-export default class extends Filter {
-
-    /**
-     * For country the final value is an array of country code strings
-     *
-     * @return {Object}
-     */
-    get values() {
-        return this.rawValues;
-    }
-}
