--- user/templates/local/participantsfilter/autocomplete_layout.mustache
@@ -1,38 +1,0 @@
-{{!
-    This file is part of Moodle - http://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-}}
-{{!
-    @template core_user/local/participantsfilter/autocomplete_layout
-
-    Moodle template for the layout of autocomplete elements.
-
-    Classes required for JS:
-    * none
-
-    Data attributes required for JS:
-    * data-region="form_autocomplete-input"
-    * data-region="form_autocomplete-suggestions"
-    * data-region="form_autocomplete-selection"
-
-    Context variables required for this template:
-    * none
-
-    Example context (json):
-    {}
-}}
-<div data-region="form_autocomplete-input"></div>
-<div data-region="form_autocomplete-suggestions"></div>
-<div data-region="form_autocomplete-selection"></div>
