--- user/templates/local/participantsfilter/autocomplete_selection.mustache
@@ -1,51 +1,0 @@
-{{!
-    This file is part of Moodle - http://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-}}
-{{!
-    @template core_user/local/participantsfilter/autocomplete_selection
-
-    Moodle template for the wrapper of currently selected items in an autocomplate form element.
-
-    Classes required for JS:
-    * none
-
-    Data attributes required for JS:
-    * none
-
-    Context variables required for this template:
-    * multiple True if this field allows multiple selections
-    * selectionId The dom id of the current selection list.
-    * items List of items with label and value fields (used by the partial).
-    * noSelectionString String to use when no items are selected (used by the partial).
-
-    Example context (json):
-    { "multiple": true, "selectionId": 1, "items": [
-        { "label": "Item label with <strong>tags</strong>", "value": "5" },
-        { "label": "Another item label with <strong>tags</strong>", "value": "4" }
-    ], "noSelectionString": "No selection" }
-}}
-<span class="sr-only" id="{{selectionId}}-label">{{#str}}selecteditems, form{{/str}}</span>
-<div{{!
-    }} class="form-autocomplete-selection d-inline-block my-0{{#multiple}} form-autocomplete-multiple h5{{/multiple}}"{{!
-    }} id="{{selectionId}}"{{!
-    }} aria-labelledby="{{selectionId}}-label"{{!
-    }} role="listbox"{{!
-    }} aria-atomic="true"{{!
-    }} tabindex="0"{{!
-    }}{{#multiple}} aria-multiselectable="true"{{/multiple}}{{!
-    }}>
-    {{> core/form_autocomplete_selection_items }}
-</div>
