--- user/templates/local/participantsfilter/filtertype.mustache
@@ -1,61 +1,0 @@
-{{!
-    This file is part of Moodle - http://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-}}
-{{!
-    @template core_user/local/participantsfilter/filtertype
-
-    Filter type data, not shown to users but used as a source of data for form autocompletion.
-
-    Classes required for JS:
-    * none
-
-    Data attributes required for JS:
-    * none
-
-    Context variables required for this template:
-    * filtertypes
-
-    Example context (json):
-    {
-        "name": "status",
-        "title": "Enrolment Status",
-        "allowcustom": "0",
-        "allowmultiple" false,
-        "filtertypeclass": "core_user/local/participantsfilter/filtertypes/courseid",
-        "values": [
-            {
-                "value": "0",
-                "title": "Inactive"
-            },
-            {
-                "value": "1",
-                "title": "Active"
-            }
-        ]
-    }
-}}
-<select {{!
-    }}{{#allowmultiple}}multiple="multiple"{{/allowmultiple}} {{!
-    }}data-field-name="{{name}}" {{!
-    }}data-field-title="{{title}}" {{!
-    }}data-allow-custom="{{allowcustom}}" {{!
-    }}class="hidden" {{!
-    }}{{#filtertypeclass}}data-filter-type-class="{{filtertypeclass}}" {{/filtertypeclass}}{{!
-}}>
-    {{#values}}
-        <option value="{{value}}">{{title}}</option>
-    {{/values}}
-</select>
