--- user/templates/local/participantsfilter/filtertypes.mustache
@@ -1,64 +1,0 @@
-{{!
-    This file is part of Moodle - http://moodle.org/
-
-    Moodle is free software: you can redistribute it and/or modify
-    it under the terms of the GNU General Public License as published by
-    the Free Software Foundation, either version 3 of the License, or
-    (at your option) any later version.
-
-    Moodle is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-    GNU General Public License for more details.
-
-    You should have received a copy of the GNU General Public License
-    along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
-}}
-{{!
-    @template core_user/local/participantsfilter/filtertypes
-
-    Placeholder to fetch all filter types.
-
-    Classes required for JS:
-    * none
-
-    Data attributes required for JS:
-    * data-filterregion="filtertypedata"
-
-    Context variables required for this template:
-    * filtertypes
-
-    Example context (json):
-    {
-        "filtertypes": [
-            {
-                "name": "status",
-                "title": "Enrolment Status",
-                "allowcustom": "0",
-                "values": [
-                    {
-                        "value": "0",
-                        "title": "Inactive"
-                    },
-                    {
-                        "value": "1",
-                        "title": "Active"
-                    }
-                ]
-            }
-        ]
-    }
-}}
-<div class="hidden" data-filterregion="filtertypedata">
-    {{#filtertypes}}
-        {{> core_user/local/participantsfilter/filtertype}}
-    {{/filtertypes}}
-</div>
-<div class="hidden">
-    <select disabled="disabled" data-filterfield="type" data-filterregion="filtertypelist">
-        <option value="">{{#str}}selectfiltertype, core_user{{/str}}</option>
-        {{#filtertypes}}
-        <option value="{{name}}">{{title}}</option>
-        {{/filtertypes}}
-    </select>
-</div>
